TOREMOVE: wifi portal to setup network on bootstrap
We may not need that in the end. Keeping this commit just in case.

Anyway what's missing:

* / goes to index.html instead of 404;
* improve detection of network issue to not start this process
  if not needed.

About this last point, if the raspberry is connected to a wired
network, the time it takes to set the network up is too long so
the ap mode is activated for the user to select a wifi which is
not necessary in that case.