fix: identity and permission issues
Local users do not exist on the remote machine so we add
`--no-same-owner` otherwise it just die on `/boot/firmware/config.txt`.

tar mode preservation (`-p`) is only needed on the target, not the source.

The script overides user's home directory file which breaks ownership and
permission of the home directory which breaks ssh auth with key so we
restore ownership just after having synced files.